updates per request
