+++ sdk/tables/azure-data-tables/inc/azure/data/tables.hpp
@@ -9,11 +9,13 @@
 #pragma once
 
 #include "azure/data/tables/account_sas_builder.hpp"
+#include "azure/data/tables/constants.hpp"
 #include "azure/data/tables/credentials/azure_sas_credential.hpp"
 #include "azure/data/tables/credentials/named_key_credential.hpp"
 #include "azure/data/tables/dll_import_export.hpp"
 #include "azure/data/tables/enum_operators.hpp"
 #include "azure/data/tables/models.hpp"
 #include "azure/data/tables/rtti.hpp"
-#include "azure/data/tables/tables_clients.hpp"
+#include "azure/data/tables/table_client.hpp"
+#include "azure/data/tables/table_service_client.hpp"
 #include "azure/data/tables/tables_sas_builder.hpp"
