+++ sdk/tables/azure-data-tables/inc/azure/data/tables/table_client.hpp
@@ -3,9 +3,11 @@
 
 #pragma once
 
+#include "azure/data/tables/constants.hpp"
 #include "azure/data/tables/credentials/azure_sas_credential.hpp"
 #include "azure/data/tables/credentials/named_key_credential.hpp"
 #include "azure/data/tables/models.hpp"
+#include "azure/data/tables/table_client_options.hpp"
 
 #include <azure/core/credentials/credentials.hpp>
 #include <azure/core/http/http.hpp>
@@ -36,85 +38,6 @@ namespace Azure { namespace Data { namespace Test {
 #endif
 
 namespace Azure { namespace Data { namespace Tables {
-
-  namespace _detail {
-    /**
-     * The package name of the SDK.
-     */
-    constexpr static const char* TablesServicePackageName = "data-tables";
-    // various strings used in the library
-    constexpr static const char* OriginHeader = "Origin";
-    constexpr static const char* AccessControlRequestMethodHeader = "Access-Control-Request-Method";
-    constexpr static const char* ResrouceTypeService = "service";
-    constexpr static const char* ComponentProperties = "properties";
-    constexpr static const char* ContentTypeXml = "application/xml";
-    constexpr static const char* ContentTypeJson = "application/json";
-    constexpr static const char* ResourceTypeHeader = "restype";
-    constexpr static const char* CompHeader = "comp";
-    constexpr static const char* ContentTypeHeader = "Content-Type";
-    constexpr static const char* ContentLengthHeader = "Content-Length";
-    constexpr static const char* AcceptHeader = "Accept";
-    constexpr static const char* PreferHeader = "Prefer";
-    constexpr static const char* PreferNoContent = "return-no-content";
-    constexpr static const char* AcceptFullMeta = "application/json;odata=fullmetadata";
-    constexpr static const char* IfMatch = "If-Match";
-    constexpr static const char* PartitionKeyFragment = "(PartitionKey='";
-    constexpr static const char* RowKeyFragment = "',RowKey='";
-    constexpr static const char* ClosingFragment = "')";
-    constexpr static const char* Value = "value";
-    constexpr static const char* TableName = "TableName";
-    constexpr static const char* ODataEditLink = "odata.editLink";
-    constexpr static const char* ODataId = "odata.id";
-    constexpr static const char* ODataType = "odata.type";
-    constexpr static const char* ODataMeta = "odata.metadata";
-    constexpr static const char* ODataError = "odata.error";
-  } // namespace _detail
-
-  /**
-   * @brief Audiences available for Blobs
-   *
-   */
-  class TablesAudience final
-      : public Azure::Core::_internal::ExtendableEnumeration<TablesAudience> {
-  public:
-    /**
-     * @brief Construct a new TablesAudience object
-     *
-     * @param tablesAudience The Azure Active Directory audience to use when forming authorization
-     * scopes. For the Language service, this value corresponds to a URL that identifies the Azure
-     * cloud where the resource is located. For more information: See
-     * https://learn.microsoft.com/en-us/azure/storage/blobs/authorize-access-azure-active-directory
-     */
-    explicit TablesAudience(std::string tablesAudience)
-        : ExtendableEnumeration(std::move(tablesAudience))
-    {
-    }
-  };
-
-  /**
-   * @brief Optional parameters for constructing a new TableClient.
-   */
-  struct TableClientOptions final : Azure::Core::_internal::ClientOptions
-  {
-    /**
-     * API version used by this client.
-     */
-    std::string ApiVersion{"2019-02-02"};
-
-    /**
-     * Enables tenant discovery through the authorization challenge when the client is configured to
-     * use a TokenCredential. When enabled, the client will attempt an initial un-authorized request
-     * to prompt a challenge in order to discover the correct tenant for the resource.
-     */
-    bool EnableTenantDiscovery = false;
-
-    /**
-     * The Audience to use for authentication with Azure Active Directory (AAD).
-     *
-     */
-    Azure::Nullable<TablesAudience> Audience;
-  };
-
   /**
    * @brief Table Client
    */
@@ -336,174 +259,5 @@ namespace Azure { namespace Data { namespace Tables {
     std::shared_ptr<Core::Http::_internal::HttpPipeline> m_pipeline;
     Core::Url m_url;
     std::string m_tableName;
-  };
-
-  /**
-   * @brief Table Service Client
-   */
-  class TableServiceClient final {
-  public:
-    /**
-     * @brief Initializes a new instance of tableServiceClient.
-     *
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     */
-    explicit TableServiceClient(const TableClientOptions& options = {});
-
-    /**
-     * @brief Initializes a new instance of tableClient.
-     *
-     * @param serviceUrl A url referencing the table that includes the name of the account and the
-     * name of the table.
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     */
-    explicit TableServiceClient(
-        const std::string& serviceUrl,
-        const TableClientOptions& options = {});
-
-    /**
-     * @brief Initializes a new instance of tableClient.
-     *
-     * @param serviceUrl A url referencing the table that includes the name of the account and the
-     * name of the table.
-     * @param credential The shared key credential used to sign requests.
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     */
-    explicit TableServiceClient(
-        const std::string& serviceUrl,
-        std::shared_ptr<Core::Credentials::TokenCredential> credential,
-        const TableClientOptions& options = {});
-
-    /**
-     * @brief Initializes a new instance of tableClient.
-     *
-     * @param serviceUrl A url referencing the table that includes the name of the account and the
-     * name of the table.
-     * @param credential The named key credential used to sign requests.
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     */
-    explicit TableServiceClient(
-        const std::string& serviceUrl,
-        std::shared_ptr<Azure::Data::Tables::Credentials::NamedKeyCredential> credential,
-        const TableClientOptions& options = {});
-
-    /**
-     * @brief Initializes a new instance of tableClient.
-     *
-     * @param serviceUrl A url referencing the table that includes the name of the account and the
-     * name of the table.
-     * @param credential The SAS credential used to sign requests.
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     */
-    explicit TableServiceClient(
-        const std::string& serviceUrl,
-        std::shared_ptr<Azure::Data::Tables::Credentials::AzureSasCredential> credential,
-        const TableClientOptions& options = {});
-
-    /**
-     * @brief Initializes a new instance of tableClient.
-     *
-     * @param connectionString the connection string used to initialize.
-     * @param options Optional client options that define the transport pipeline policies for
-     * authentication, retries, etc., that are applied to every request.
-     * @return TableServiceClient.
-     */
-    static TableServiceClient CreateFromConnectionString(
-        const std::string& connectionString,
-        const TableClientOptions& options = {});
-
-    /**
-     * @brief Creates a new table under the given account.
-     *
-     * @param context for canceling long running operations.
-     * @param tableName The name of the table to be created.
-     * @return Create table result.
-     */
-    Response<Models::Table> CreateTable(
-        std::string const& tableName,
-        Core::Context const& context = {});
-
-    /**
-     * @brief Operation permanently deletes the specified table.
-     *
-     * @param context for canceling long running operations.
-     * @param tableName The name of the table to be deleted.
-     * @return Delete table result.
-     */
-    Response<Models::DeleteTableResult> DeleteTable(
-        std::string const& tableName,
-        Core::Context const& context = {});
-
-    /**
-     * @brief Queries tables under the given account.
-     *
-     * @param options Optional parameters to execute this function.
-     * @param context for canceling long running operations.
-     * @return Query tables paged response.
-     */
-    Models::QueryTablesPagedResponse QueryTables(
-        const Models::QueryTablesOptions& options = {},
-        const Azure::Core::Context& context = {}) const;
-
-    /**
-     * @brief Set service properties
-     *
-     * @param options Optional parameters to execute this function.
-     * @param context for canceling long running operations.
-     */
-    Response<Models::SetServicePropertiesResult> SetServiceProperties(
-        Models::SetServicePropertiesOptions const& options = {},
-        Core::Context const& context = {});
-
-    /**
-     * @brief Get service properties
-     *
-     * @param context for canceling long running operations.
-     * @return Get service properties result.
-     */
-    Response<Models::TableServiceProperties> GetServiceProperties(
-        Core::Context const& context = {});
-
-    /**
-     * @brief Get service statistics
-     *
-     * @param context for canceling long running operations.
-     * @return Get service statistics result.
-     */
-    Response<Models::ServiceStatistics> GetStatistics(Core::Context const& context = {});
-
-    /**
-     * @brief Pre flight check
-     *
-     * @param options Optional parameters to execute this function.
-     * @param context for canceling long running operations.
-     * @return Get service statistics result.
-     */
-    Response<Models::PreflightCheckResult> PreflightCheck(
-        Models::PreflightCheckOptions const& options,
-        Core::Context const& context = {});
-    /**
-     * @brief Get table client.
-     *
-     * @param tableName The name of the table.
-     * @param options Optional parameters for the table client.
-     * @return TableClient.
-     * @remark The TableClient can be used to perform operations on the table. This method will
-     * attempt to create a table client with the same credentials as the service client except the
-     * SAS token credential as the SAS token permissions varies from the service level permissions.
-     */
-    TableClient GetTableClient(const std::string& tableName, TableClientOptions const& options = {})
-        const;
-
-  private:
-    std::shared_ptr<Core::Http::_internal::HttpPipeline> m_pipeline;
-    std::shared_ptr<Core::Credentials::TokenCredential> m_tokenCredential;
-    std::shared_ptr<Azure::Data::Tables::Credentials::NamedKeyCredential> m_namedKeyCredential;
-    Core::Url m_url;
   };
 }}} // namespace Azure::Data::Tables
